Split downlaoder and network manager
